Removed unneccesary files, refactored to match other Filamnet components
