+++ src/Columns/Components/RatingStar.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace IbrahimBougaoua\FilamentRatingStar\Columns;
+namespace IbrahimBougaoua\FilamentRatingStar\Columns\Components;
 
 use Filament\Forms\Components\Concerns\HasToggleColors;
 use Filament\Forms\Components\Concerns\HasToggleIcons;
@@ -9,14 +9,14 @@ use Filament\Tables\Columns\Concerns\CanBeValidated;
 use Filament\Tables\Columns\Concerns\CanUpdateState;
 use Filament\Tables\Columns\Contracts\Editable;
 
-class RatingStarColumn extends Column implements Editable
+class RatingStar extends Column implements Editable
 {
     use CanBeValidated;
     use CanUpdateState;
     use HasToggleColors;
     use HasToggleIcons;
 
-    protected string $view = 'filament-rating-star::columns.rating-star-column';
+    protected string $view = 'filament-rating-star::columns.components.rating-star';
 
     protected string $size = 'lg';
