--- src/FilamentRatingStar.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace IbrahimBougaoua\FilamentRatingStar;
-
-class FilamentRatingStar
-{
-}
