--- workbench/app/Providers/WorkbenchServiceProvider.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace Workbench\App\Providers;
-
-use Illuminate\Support\Facades\Route;
-use Illuminate\Support\ServiceProvider;
-
-class WorkbenchServiceProvider extends ServiceProvider
-{
-    /**
-     * Register services.
-     */
-    public function register(): void
-    {
-        //
-    }
-
-    /**
-     * Bootstrap services.
-     */
-    public function boot(): void
-    {
-        Route::view('/', 'welcome');
-    }
-}
